reorganized the components ordener
