+++ components/Styling/Card.js
