+++ components/Styling/Main.js
