+++ pages/lists/placelist.js
@@ -1,4 +1,4 @@
-import ShowPlaceList from "@/components/ShowPlaceList";
+import ShowPlaceList from "@/components/Places/ShowPlaceList";
 
 export default function PlaceList({ places }) {
   return <ShowPlaceList />;
